renamed file protocol_wrapper.rs to json_protocol.rs plus fixed test
