+++ net_connection/src/json_protocol.rs
