+++ test_bin/src/main.rs
@@ -13,7 +13,7 @@ pub mod p2p_node;
 use holochain_core_types::cas::content::Address;
 use holochain_net_connection::{
     net_connection::NetSend,
-    protocol_wrapper::{
+    json_protocol::{
         ConnectData, DhtData, DhtMetaData, GetDhtData, GetDhtMetaData, JsonProtocol, MessageData,
         TrackDnaData,
     },
